moved PersonInput and PersonBirthInput to resolvers/types.ts
